# docs/swagger-old.yaml
openapi: 3.0.0
info:
  title: APIServiceGo
  version: 1.0.0
  description: EV Charging API for searching charging connectors, getting estimates, managing orders, and processing payments for electric vehicle charging sessions.
servers:
  - url: http://{{localhost}}
paths:
  /v1/search:
    post:
      operationId: searchChargingConnectors
      summary: Search Charging Connectors
      tags:
        - Discovery
      parameters:
        - name: page
          in: query
          description: Page number for pagination.
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          description: Page size for pagination.
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: OK
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
              example:
                total: 1
                page: 1
                per_page: 100
                catalogs:
                  - id: catalog-ev-charging-001
                    provider:
                      id: ecopower-charging
                      descriptor:
                        name: EcoPower Charging Pvt Ltd
                      address:
                        name: MG JVLR Jogeshwari Caves Road
                        geo_coordinates: [12.9716, 77.5946]
                    rating:
                      value: 4.5
                      count: 128
                    connectors:
                      - id: ev-charger-ccs2-001
                        isActive: true
                        availabilityWindow:
                          - startTime: "06:00:00"
                            endTime: "22:00:00"
                        price:
                          value: "8"
                          currency: "INR/kWH"
                        connectorAttributes:
                          connectorType: TYPE_2
                          maxPowerKW: 60
                          minPowerKW: 5
                          socketCount: 2
                          reservationSupported: true
                          status: Available
                          chargingSpeed: FAST
                          powerType: DC
                          connectorFormat: CABLE
                    offers:
                      - id: offer-ccs2-60kw-kwh
                        descriptor:
                          name: Per-kWh Tariff - CCS2 60kW
                        items:
                          - ev-charger-ccs2-001
                        price:
                          currency: INR
                          value: 18
                          applicableQuantity:
                            unitText: Kilowatt Hour
                            unitCode: KWH
                            unitQuantity: 1
                        validity:
                          startDate: "2025-10-01T00:00:00Z"
                          endDate: "2026-03-31T23:59:59Z"
                        acceptedPaymentMethod:
                          - UPI
                          - Card
                          - Wallet
                        offerAttributes:
                          buyerFinderFee:
                            feeType: PERCENTAGE
                            feeValue: 2.5
                          idleFeePolicy: ₹2/min after 10 min post-charge
                        provider: ecopower-charging
                      - id: offer-ccs2-120kw-kwh
                        descriptor:
                          name: Per-kWh Tariff - CCS2 120kW
                        items:
                          - ev-charger-ccs2-002
                        price:
                          currency: INR
                          value: 22
                          applicableQuantity:
                            unitText: Kilowatt Hour
                            unitCode: KWH
                            unitQuantity: 1
                        validity:
                          startDate: "2025-10-15T00:00:00Z"
                          endDate: "2026-04-15T23:59:59Z"
                        acceptedPaymentMethod:
                          - UPI
                          - Card
                          - Wallet
                          - NetBanking
                        offerAttributes:
                          buyerFinderFee:
                            feeType: PERCENTAGE
                            feeValue: 2
                          idleFeePolicy: ₹3/min after 15 min post-charge
                        provider: greencharge-koramangala
                      - id: offer-type2-22kw-kwh
                        descriptor:
                          name: Per-kWh Tariff - Type 2 22kW
                        items:
                          - ev-charger-type2-001
                        price:
                          currency: INR
                          value: 15
                          applicableQuantity:
                            unitText: Kilowatt Hour
                            unitCode: KWH
                            unitQuantity: 1
                        validity:
                          startDate: "2025-10-20T00:00:00Z"
                          endDate: "2026-04-20T23:59:59Z"
                        acceptedPaymentMethod:
                          - UPI
                          - Card
                          - Wallet
                        offerAttributes:
                          buyerFinderFee:
                            feeType: PERCENTAGE
                            feeValue: 1.5
                          idleFeePolicy: ₹1/min after 30 min post-charge
                        provider: powergrid-indiranagar
        "400":
          description: Bad Request - Invalid request parameters or malformed request body.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "BAD_REQUEST"
                  message: "Invalid request parameters or malformed request body."
        "401":
          description: Unauthorized - Invalid or missing authentication token.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or missing authentication token."
        "422":
          description: Unprocessable Entity - Validation error in request data.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "VALIDATION_ERROR"
                  message: "Validation error in request data."
        "500":
          description: Internal Server Error - Server error occurred while processing the request.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "INTERNAL_ERROR"
                  message: "Server error occurred while processing the request."
      description: >-
        Search for available charging connectors and offers near a location or
        by filters. Either evse_id or geo_coordinates with distance_meters is required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
  /v1/estimate:
    post:
      operationId: getEstimates
      summary: Get Estimates
      tags:
        - Discovery
      parameters:
        - name: X-Transaction-Id
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateResponse"
              example:
                order:
                  id: "1231208-id"
                  mode: reservation
                  status: quoted_price
                amount:
                  value: 128.64
                  currency: INR
                durationInMinutes: "15"
                percentageOfBatteryCharged: "80"
                energy: 
                  value: 30
                  unit: kWh
                validity:
                  startDate: "2025-01-27T00:00:00Z"
                  endDate: "2025-04-27T23:59:59Z"
                priceComponents:
                  - type: UNIT
                    value: 100
                    currency: INR
                    description: Base charging session cost (100 INR)
                  - type: SURCHARGE
                    value: 20
                    currency: INR
                    description: Surge price (20%)
                  - type: DISCOUNT
                    value: -15
                    currency: INR
                    description: Offer discount (15%)
                  - type: FEE
                    value: 10
                    currency: INR
                    description: Service fee
                  - type: FEE
                    value: 13.64
                    currency: INR
                    description: Overcharge estimation
                  - type: "Pending Payment"
                    value: 0.64
                    currency: INR
                    description: Pending payment
                cancellation:
                  fee:
                    percentage: "30"
                  externalRef:
                    mimetype: text/html
                    url: https://example-company.com/charge/tnc.html
        "400":
          description: Bad Request - Invalid request parameters or missing required fields.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "BAD_REQUEST"
                  message: "Invalid request parameters or missing required fields."
        "401":
          description: Unauthorized - Invalid or missing authentication token.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or missing authentication token."
        "404":
          description: Not Found - EVSE or connector not found.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "NOT_FOUND"
                  message: "EVSE or connector not found."
        "422":
          description: Unprocessable Entity - Validation error in request data.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "VALIDATION_ERROR"
                  message: "Validation error in request data."
        "500":
          description: Internal Server Error - Server error occurred while processing the request.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "INTERNAL_ERROR"
                  message: "Server error occurred while processing the request."
      description: Get estimates for a prospective charging session and offer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateRequest"
  /v1/orders/{order_id}/payment:
    post:
      operationId: initiatePayment
      summary: Initiate Payment
      tags:
        - Payments
      parameters:
        - name: X-Transaction-Id
          in: header
          required: true
          schema:
            type: string
        - name: X-Bpp-Id
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
              example:
                order:
                  id: order-bpp-789012
                  status: ACTIVE
                  mode: RESERVATION
                amount:
                  value: 128.64
                  currency: INR
                beneficiaryId: ""
                acceptedPaymentMethod:
                  - BankTransfer
                  - UPI
                  - Wallet
                paymentUrl: ""
                validity:
                  startDate: "2025-01-27T00:00:00Z"
                  endDate: "2025-04-27T23:59:59Z"
        "400":
          description: Bad Request - Invalid request parameters or malformed request body.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "BAD_REQUEST"
                  message: "Invalid request parameters or malformed request body."
        "401":
          description: Unauthorized - Invalid or missing authentication token.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or missing authentication token."
        "404":
          description: Not Found - Order not found.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "NOT_FOUND"
                  message: "Order not found."
        "409":
          description: Conflict - Payment already processed or order in invalid state for payment.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "CONFLICT"
                  message: "Payment already processed or order in invalid state for payment."
        "422":
          description: Unprocessable Entity - Validation error in payment data.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "VALIDATION_ERROR"
                  message: "Validation error in payment data."
        "500":
          description: Internal Server Error - Server error occurred while processing the request.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "INTERNAL_ERROR"
                  message: "Server error occurred while processing the request."
      description: Initiate or fetch payment instructions for the specified order.
      requestBody:
        content:
          application/json:
            schema:
              type: object
    parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          default: <string>
  /v1/orders/{order_id}:
    get:
      summary: get order info
      tags:
        - Orders
      parameters:
        - name: X-Transaction-Id
          in: header
          required: true
          schema:
            type: string
        - name: X-Bpp-Id
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                default: <string>
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
              example:
                order:
                  id: order-bpp-789012
                  status: ACTIVE
                  mode: RESERVATION
                payment:
                  status: PAID
                charging:
                  status: ACTIVE
                connectorId: ""
                connectorType: ""
                vehicle:
                  type: ""
                  model: ""
                  make: ""
                trackingUrl: https://track.bluechargenet-aggregator.io/session/SESSION-9876543210
                chargingTelemetry:
                  eventTime: "2025-01-27T17:00:00Z"
                  metrics:
                    - name: STATE_OF_CHARGE
                      value: 62.5
                      unitCode: PERCENTAGE
                    - name: POWER
                      value: 18.4
                      unitCode: KWH
                    - name: ENERGY
                      value: 10.2
                      unitCode: KW
                    - name: VOLTAGE
                      value: 392
                      unitCode: VLT
                    - name: CURRENT
                      value: 47
                      unitCode: AMP
                    - name: SESSION_DURATION
                      value: 10
                      unitCode: min
        "401":
          description: Unauthorized - Invalid or missing authentication token.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "UNAUTHORIZED"
                  message: "Invalid or missing authentication token."
        "404":
          description: Not Found - Order not found.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "NOT_FOUND"
                  message: "Order not found."
        "500":
          description: Internal Server Error - Server error occurred while processing the request.
          headers:
            X-Transaction-Id:
              required: true
              schema:
                type: string
                description: Transaction identifier for request tracking
            X-Bpp-Id:
              required: true
              schema:
                type: string
                description: BPP identifier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: "INTERNAL_ERROR"
                  message: "Server error occurred while processing the request."
      description: >-
        Retrieve order details, current status, and telemetry for the specified
        order.
    parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
...
components:
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code
              example: "INVALID_REQUEST"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid request parameters"
            details:
              type: object
              description: Additional error details
              additionalProperties: true
...
    SearchResponse:
      type: object
      required:
        - total
        - page
        - per_page
        - catalogs
      properties:
        total:
          type: integer
          description: Total number of catalogs matching the search criteria
        page:
          type: integer
          description: Current page number
        per_page:
          type: integer
          description: Number of items per page
        catalogs:
          type: array
          description: List of catalogs containing connectors and offers
          items:
            $ref: "#/components/schemas/Catalog"
...
    TimeWindow:
      type: object
      description: Desired time window
      properties:
        start:
          type: string
          format: date-time
          description: Start date and time
        end:
          type: string
          format: date-time
          description: End date and time
    SearchRequest:
      type: object
      properties:
        evse_id:
          type: string
          description: Specific EVSE identifier to look up
        geo_coordinates:
          type: array
          minItems: 2
          maxItems: 2
          description: Geographic coordinates as [latitude, longitude] array
          items:
            type: number
          example: [12.9716, 77.5946]
        distance_meters:
          type: number
          format: float
          minimum: 0
          maximum: 50000
          example: 5000
          description: Radius of search in meters (0-50000)
        time_window:
          $ref: "#/components/schemas/TimeWindow"
        filters:
          $ref: "#/components/schemas/SearchFilters"
        sort:
          $ref: "#/components/schemas/SearchSort"
      oneOf:
        - type: object
          required:
            - evse_id
          properties:
            evse_id:
              type: string
              description: Specific EVSE identifier to look up
            time_window:
              $ref: "#/components/schemas/TimeWindow"
            filters:
              $ref: "#/components/schemas/SearchFilters"
        - type: object
          required:
            - geo_coordinates
            - distance_meters
          properties:
            geo_coordinates:
              type: array
              minItems: 2
              maxItems: 2
              description: Geographic coordinates as [latitude, longitude] array
              items:
                type: number
              example: [12.9716, 77.5946]
            distance_meters:
              type: number
              format: float
              minimum: 0
              maximum: 50000
              example: 5000
              description: Radius of search in meters (0-50000)
            time_window:
              $ref: "#/components/schemas/TimeWindow"
            filters:
              $ref: "#/components/schemas/SearchFilters"
...
    EstimateRequest:
      type: object
      required:
        - evse_id
        - vehicle
        - connector_id
      properties:
        evse_id:
          type: string
          description: EVSE identifier of the selected connector/station
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        connector_id:
          type: string
        time_window:
          $ref: "#/components/schemas/TimeWindow"
        energy:
          $ref: "#/components/schemas/Energy"
        offer_id:
          type: string
        amount:
          $ref: "#/components/schemas/Amount"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
security:
  - BearerAuth: []
...